fix: выделить абстактный тестовый класс, заменить дубли сообщений об ошибках на константы
