+++ test/FileBackedTaskManagerTest.java
@@ -1,7 +1,6 @@
 import managers.tasks.FileBackedTaskManager;
 import managers.tasks.ManagerSaveException;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tasks.Epic;
 import tasks.Subtask;
@@ -23,12 +22,22 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class FileBackedTaskManagerTest {
+class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager> {
     private static final String CURRENT_DIR = System.getProperty("user.dir");
     private Path testDB;
+    private final Path defaultDB = Paths.get(CURRENT_DIR, "db.csv");
 
-    @BeforeEach
-    void beforeEach() {
+    @Override
+    protected FileBackedTaskManager createTaskManager() {
+        return new FileBackedTaskManager();
+    }
+
+    @Override
+    void addTasks() {
+        //Вызов в методе @BeforeEach абстрактоного класса
+        //Для проверки дефолтной версии FileBackedTaskManager (db.csv)
+        super.addTasks();
+        //Для проверки версии загруженной из готового файла
         try {
             testDB = Files.createTempFile(Paths.get(CURRENT_DIR), "testDB", ".csv");
             try (Writer writer = new FileWriter(testDB.toFile())) {
@@ -51,107 +60,58 @@ class FileBackedTaskManagerTest {
     void afterEach() {
         try {
             Files.delete(testDB);
+            Files.delete(defaultDB);
         } catch (IOException e) {
             throw new ManagerSaveException(e.getMessage(), e.getCause());
         }
     }
 
     @Test
-    void createDefaultManager() {
-        FileBackedTaskManager fileBackedTaskManager = new FileBackedTaskManager();
-        fileBackedTaskManager.createTask(new Task("Task1", "First task to complete", Status.NEW,
-                LocalDateTime.of(2024, 5, 12, 18, 30), Duration.ofDays(2)));
-        String fileName = "db.csv";
-        Path file = Paths.get(CURRENT_DIR, fileName);
-        assertTrue(Files.exists(file), "Файл по умолчанию не создан");
-
-        fileBackedTaskManager.createTask(new Epic("Epic1", "Epic to complete"));
-        fileBackedTaskManager.createTask(new Subtask("Subtask1", "First Subtask to first epic",
-                Status.DONE, 2, LocalDateTime.of(2024, 4, 28, 9, 0),
-                Duration.ofHours(8)));
-        fileBackedTaskManager.createTask(new Subtask("Subtask2", "Second Subtask to first epic",
-                Status.NEW, 2, LocalDateTime.of(2024, 5, 15, 10, 30),
-                Duration.ofMinutes(25)));
-        fileBackedTaskManager.getTask(1);
-        fileBackedTaskManager.getTask(2);
-        fileBackedTaskManager.getTask(1);
-
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(file.toFile());
-        List<Task> history = afterLoadFileBackedTaskManager.getHistory();
-        assertNotNull(history, "История не возвращается из managers.history.HistoryManager");
-        System.out.println(afterLoadFileBackedTaskManager.getTask(1));
-        assertEquals(2, history.size(), "Возвращается некорректный список истории просмотров задач");
-        assertEquals("Task{id='1, 'name='Task1', description='First task to complete', status=NEW'}",
-                afterLoadFileBackedTaskManager.getTask(1).toString(), "Созданная и полученная задача " +
-                        "не совпадает");
-        assertEquals("Epic{id='2, 'name='Epic1', description='Epic to complete', status=IN_PROGRESS, "
-                        + "subtasksNumber='2'}", afterLoadFileBackedTaskManager.getTask(2).toString(),
-                "Созданный и полученный эпик не совпадает");
-        assertEquals("Subtask{id='3, 'name='Subtask1', description='First Subtask to first epic', "
-                        + "status=DONE, epicID='2'}", afterLoadFileBackedTaskManager.getTask(3).toString(),
-                "Созданная и полученная подзадача не совпадает");
-        assertEquals("Subtask{id='4, 'name='Subtask2', description='Second Subtask to first epic', "
-                        + "status=NEW, epicID='2'}", afterLoadFileBackedTaskManager.getTask(4).toString(),
-                "Созданная и полученная подзадача не совпадает");
-        try {
-            Files.delete(file);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    @Test
     void loadManager() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
 
-        assertEquals(1, fileBackedTaskManager.getTasksList().size(),
-                "Некорректный размер пулла задач");
-        assertEquals(2, fileBackedTaskManager.getEpicsList().size(),
-                "Некорректный размер пулла эпиков");
-        assertEquals(3, fileBackedTaskManager.getSubtasksList().size(),
-                "Некорректный размер пулла подзадач");
+        assertEquals(1, taskManager.getTasksList().size(), taskPoolSizeError);
+        assertEquals(2, taskManager.getEpicsList().size(), epicPoolSizeError);
+        assertEquals(3, taskManager.getSubtasksList().size(), subtaskPoolSizeError);
 
-        List<Task> history = fileBackedTaskManager.getHistory();
-        assertNotNull(history, "История не возвращается из managers.history.HistoryManager");
-        assertEquals(2, history.size(), "Возвращается некорректный список истории просмотров задач");
+        List<Task> history = taskManager.getHistory();
+        assertNotNull(history, historyReturnNull);
+        assertEquals(2, history.size(), historySizeError);
 
         assertEquals("Task{id='1, 'name='Task1', description='First task to complete', status=NEW'}",
-                fileBackedTaskManager.getTask(1).toString(), "Созданная и полученная задача не совпадает");
+                taskManager.getTask(1).toString(), taskInDBNotEqualToCreated);
         assertEquals("Epic{id='2, 'name='Epic1', description='First epic to complete', status=NEW, "
-                        + "subtasksNumber='2'}", fileBackedTaskManager.getTask(2).toString(),
-                "Созданный и полученный эпик не совпадает");
+                        + "subtasksNumber='2'}", taskManager.getTask(2).toString(),
+                taskInDBNotEqualToCreated);
         assertEquals("Subtask{id='6, 'name='Subtask3', description='First subtask to second epic', "
-                        + "status=NEW, epicID='3'}", fileBackedTaskManager.getTask(6).toString(),
-                "Созданная и полученная подзадача не совпадает");
+                        + "status=NEW, epicID='3'}", taskManager.getTask(6).toString(),
+                taskInDBNotEqualToCreated);
     }
 
     @Test
     void removeTaskPool() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeTaskPool();
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertTrue(afterLoadFileBackedTaskManager.getTasksList().isEmpty(), "Некорректный размер пулла задач");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeTaskPool();
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertTrue(taskManager.getTasksList().isEmpty(), taskPoolSizeError);
     }
 
     @Test
     void removeEpicPool() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeEpicPool();
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertTrue(afterLoadFileBackedTaskManager.getEpicsList().isEmpty(),
-                "Некорректный размер пулла эпиков");
-        assertTrue(afterLoadFileBackedTaskManager.getSubtasksList().isEmpty(),
-                "Некорректный размер пулла подзадач");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeEpicPool();
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertTrue(taskManager.getEpicsList().isEmpty(),epicPoolSizeError);
+        assertTrue(taskManager.getSubtasksList().isEmpty(), subtaskPoolSizeError);
     }
 
     @Test
     void removeSubtaskPool() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeSubtaskPool();
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertTrue(afterLoadFileBackedTaskManager.getSubtasksList().isEmpty(),
-                "Некорректный размер пулла подзадач");
-        List<Epic> epics = afterLoadFileBackedTaskManager.getEpicsList();
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeSubtaskPool();
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertTrue(taskManager.getSubtasksList().isEmpty(), subtaskPoolSizeError);
+        List<Epic> epics = taskManager.getEpicsList();
         for (Epic epic : epics) {
             assertTrue(epic.getSubTasksIds().isEmpty(), String.format("У Эпика %d не очищены подзадачи", epic.getId()));
         }
@@ -159,94 +119,87 @@ class FileBackedTaskManagerTest {
 
     @Test
     void correctHistoryOrder() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.getTask(1);
-        fileBackedTaskManager.getTask(3);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        List<Task> history = afterLoadFileBackedTaskManager.getHistory();
-        assertNotNull(history, "История не возвращается из managers.history.HistoryManager");
-        assertEquals(3, history.size(), "Возвращается некорректный список истории просмотров задач");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.getTask(1);
+        taskManager.getTask(3);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        List<Task> history = taskManager.getHistory();
+        assertNotNull(history, historyReturnNull);
+        assertEquals(3, history.size(), historySizeError);
         assertEquals("Subtask{id='6, 'name='Subtask3', description='First subtask to second epic', "
-                        + "status=NEW, epicID='3'}", history.get(0).toString(),
-                "Некорректная последовательность вывода истории");
+                        + "status=NEW, epicID='3'}", history.get(0).toString(), historyOrderError);
         assertEquals("Task{id='1, 'name='Task1', description='First task to complete', status=NEW'}",
-                history.get(1).toString(), "Некорректная последовательность вывода истории");
+                history.get(1).toString(), historyOrderError);
         assertEquals("Epic{id='3, 'name='Epic2', description='Second epic to complete', status=NEW, "
-                        + "subtasksNumber='1'}", history.get(2).toString(),
-                "Некорректная последовательность вывода истории");
+                        + "subtasksNumber='1'}", history.get(2).toString(), historyOrderError);
     }
 
     @Test
     void correctMaintainingIdNumbering() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.createTask(new Task("Task2", "Second task to complete", Status.NEW,
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.createTask(new Task("Task2", "Second task to complete", Status.NEW,
                 LocalDateTime.of(2024, 7, 28, 12, 0), Duration.ofHours(2)));
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
         assertEquals("Task{id='7, 'name='Task2', description='Second task to complete', status=NEW'}",
-                afterLoadFileBackedTaskManager.getTask(7).toString(), "Некорректная нумерация новых задач");
+                taskManager.getTask(7).toString(), "Некорректная нумерация новых задач");
     }
 
     @Test
     void updateTask() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
         Task newTask = new Task("Task1", "First task updated Status", Status.DONE,
                 LocalDateTime.of(2024, 5, 12, 18, 30), Duration.ofDays(2));
         newTask.setId(1);
-        fileBackedTaskManager.updateTask(newTask);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.updateTask(newTask);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
         assertEquals("Task{id='1, 'name='Task1', description='First task updated Status', status=DONE'}",
-                afterLoadFileBackedTaskManager.getTask(1).toString(), "Некорректно обновлена задача");
+                taskManager.getTask(1).toString(), incorrectUpdating);
     }
 
     @Test
     void updateSubtask() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
         Subtask newSubtask = new Subtask("Subtask2", "Second subtask to first epic updated",
                 Status.DONE, 2, LocalDateTime.of(2024, 6, 1, 9, 30),
                 Duration.ofDays(4));
         newSubtask.setId(5);
-        fileBackedTaskManager.updateTask(newSubtask);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.updateTask(newSubtask);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
         assertEquals("Epic{id='2, 'name='Epic1', description='First epic to complete', status=IN_PROGRESS, "
-                        + "subtasksNumber='2'}", afterLoadFileBackedTaskManager.getTask(2).toString(),
-                "У соответствующего эпика не пересмотрен статус");
+                        + "subtasksNumber='2'}", taskManager.getTask(2).toString(), epicStatusError);
         assertEquals("Subtask{id='5, 'name='Subtask2', description='Second subtask to first epic updated', "
-                        + "status=DONE, epicID='2'}", afterLoadFileBackedTaskManager.getTask(5).toString(),
-                "Некорректно обновлена задача");
+                        + "status=DONE, epicID='2'}", taskManager.getTask(5).toString(), incorrectUpdating);
     }
 
     @Test
     void removeTask() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeTask(1);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertTrue(afterLoadFileBackedTaskManager.getTasksList().isEmpty(), "Некорректный размер пулла задач");
-        List<Task> history = afterLoadFileBackedTaskManager.getHistory();
-        assertEquals(1, history.size(), "Возвращается некорректный список истории просмотров задач");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeTask(1);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertTrue(taskManager.getTasksList().isEmpty(), taskPoolSizeError);
+        List<Task> history = taskManager.getHistory();
+        assertEquals(1, history.size(), historySizeError);
     }
 
     @Test
     void removeEpic() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeTask(2);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertEquals(1, afterLoadFileBackedTaskManager.getEpicsList().size(),
-                "Некорректный размер пулла эпиков");
-        assertEquals(1, afterLoadFileBackedTaskManager.getSubtasksList().size(),
-                "Некорректный размер пулла подзадач");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeTask(2);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertEquals(1, taskManager.getEpicsList().size(), epicPoolSizeError);
+        assertEquals(1, taskManager.getSubtasksList().size(), subtaskPoolSizeError);
     }
 
     @Test
     void removeSubtask() {
-        FileBackedTaskManager fileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        fileBackedTaskManager.removeTask(6);
-        FileBackedTaskManager afterLoadFileBackedTaskManager = FileBackedTaskManager.load(testDB.toFile());
-        assertEquals(2, afterLoadFileBackedTaskManager.getSubtasksList().size(),
-                "Некорректный размер пулла подзадач");
-        List<Task> history = afterLoadFileBackedTaskManager.getHistory();
-        assertEquals(1, history.size(), "Возвращается некорректный список истории просмотров задач");
-        Epic connectedEpic = (Epic) afterLoadFileBackedTaskManager.getTask(3);
-        assertTrue(connectedEpic.getSubTasksIds().isEmpty(), "У Эпика не удалена подзадача");
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        taskManager.removeTask(6);
+        taskManager = FileBackedTaskManager.load(testDB.toFile());
+        assertEquals(2, taskManager.getSubtasksList().size(), subtaskPoolSizeError);
+        List<Task> history = taskManager.getHistory();
+        assertEquals(1, history.size(), historySizeError);
+        Epic connectedEpic = (Epic) taskManager.getTask(3);
+        assertTrue(connectedEpic.getSubTasksIds().isEmpty(), epicPoolSizeError);
     }
 
     @Test
